Split up settings
