+++ rust/hyperlane-base/src/macros.rs
@@ -118,10 +118,10 @@ macro_rules! decl_settings {
 /// Static logic called by the decl_settings! macro. Do not call directly!
 #[doc(hidden)]
 pub fn _new_settings<'de, T: Deserialize<'de>>(name: &str) -> eyre::Result<T> {
-    use crate::settings;
+    use crate::settings::loader::load_settings_object;
     use std::env;
 
-    settings::load_settings_object::<_, &str>(
+    load_settings_object::<T, &str>(
         name,
         Some(&env::var("BASE_CONFIG").unwrap_or_else(|_| "base".into())),
         &[],
