+++ rust/hyperlane-base/src/settings/signers.rs
@@ -1,0 +1,62 @@
+use crate::settings::KMS_CLIENT;
+use ethers::prelude::AwsSigner;
+use eyre::{bail, Report};
+use hyperlane_core::utils::HexString;
+use hyperlane_core::Signers;
+use rusoto_core::credential::EnvironmentProvider;
+use rusoto_core::HttpClient;
+use rusoto_kms::KmsClient;
+use tracing::instrument;
+
+/// Ethereum signer types
+#[derive(Debug, Clone, serde::Deserialize)]
+#[serde(tag = "type", rename_all = "camelCase")]
+pub enum SignerConf {
+    /// A local hex key
+    HexKey {
+        /// Hex string of private key, without 0x prefix
+        key: HexString<64>,
+    },
+    /// An AWS signer. Note that AWS credentials must be inserted into the env
+    /// separately.
+    Aws {
+        /// The UUID identifying the AWS KMS Key
+        id: String, // change to no _ so we can set by env
+        /// The AWS region
+        region: String,
+    },
+    #[serde(other)]
+    /// Assume node will sign on RPC calls
+    Node,
+}
+
+impl Default for SignerConf {
+    fn default() -> Self {
+        Self::Node
+    }
+}
+
+impl SignerConf {
+    /// Try to convert the ethereum signer to a local wallet
+    #[instrument(err)]
+    pub async fn try_into_signer(&self) -> Result<Signers, Report> {
+        match self {
+            SignerConf::HexKey { key } => Ok(Signers::Local(key.as_ref().parse()?)),
+            SignerConf::Aws { id, region } => {
+                let client = KMS_CLIENT.get_or_init(|| {
+                    KmsClient::new_with_client(
+                        rusoto_core::Client::new_with(
+                            EnvironmentProvider::default(),
+                            HttpClient::new().unwrap(),
+                        ),
+                        region.parse().expect("invalid region"),
+                    )
+                });
+
+                let signer = AwsSigner::new(client, id, 0).await?;
+                Ok(Signers::Aws(signer))
+            }
+            SignerConf::Node => bail!("Node signer"),
+        }
+    }
+}
